Fix: findable module made more universal - this commit includes all old code commented out
