+++ lib/league.rb
@@ -53,17 +53,19 @@ class League
   end
 
   def all_games_played(team_id)
-    @game_teams.select do |row|
-      row if row[:team_id] == team_id
-    end
+    # @game_teams.select do |row|
+    #   row[:team_id] == team_id
+    # end
+    find_in_sheet(team_id, :team_id, @game_teams)
   end
 
   def convert_team_id_to_name(team_id)
-    name_array = []
-    team = @teams.find do |row|
-      row[:team_id] == team_id
-    end
-    team[:teamname]
+    # name_array = []
+    # team = @teams.find do |row|
+    #   row[:team_id] == team_id
+    # end
+    team = find_in_sheet(team_id, :team_id, @teams)
+    team[0][:teamname]
   end
 
   def highest_scoring_visitor
